+++ lib/ngt/index.rb
@@ -2,7 +2,9 @@ module Ngt
   class Index
     include Utils
 
-    attr_reader :dimensions, :edge_size_for_creation, :edge_size_for_search, :path
+    DISTANCE_TYPES = ["L1", "L2", "Hamming", "Angle", "Cosine", "NormalizedAngle", "NormalizedCosine", "Jaccard"]
+
+    attr_reader :dimensions, :distance_type, :edge_size_for_creation, :edge_size_for_search, :path
 
     def initialize(path)
       @path = path
@@ -13,6 +15,9 @@ module Ngt
       ffi(:ngt_get_property, @index, property)
 
       @dimensions = ffi(:ngt_get_property_dimension, property)
+      @distance_type = DISTANCE_TYPES[ffi(:ngt_get_property_distance_type, property)]
+
+
       @edge_size_for_creation = ffi(:ngt_get_property_edge_size_for_creation, property)
       @edge_size_for_search = ffi(:ngt_get_property_edge_size_for_search, property)